LayoutTests/http/tests/security/contentSecurityPolicy/1.1/child-src/worker-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-fail.js because it does not appear in the child-src directive of the Content Security Policy.
Workers should be governed by 'child-src'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS var w = new Worker('/security/contentSecurityPolicy/resources/alert-fail.js'); threw exception Error: SecurityError: DOM Exception 18.
PASS successfullyParsed is true

TEST COMPLETE
